Simplify the code in Perl_nextargv().
Split the ternary that called Perl_do_open_raw() and Perl_do_open6() based
on PL_inplace into two different if blocks, and merge these with the following
code which is also conditional on PL_inplace.

Remove the warning code from an else block and re-indent it, to make it clear
that it is always called if control reaches the end of the while loop.