lib/locale.t: Fix broken test
The test that [:digit:] is a subset of [:xdigit:] failed in locales
where  [:digit:] matches 2 blocks of 10 digits, but the second block
isn't considered part of [:xdigit:].  This happens in Thai on Windows.
The POSIX standard http://pubs.opengroup.org/onlinepubs/9699919799/
does not use very clear language, but I'm taking it as meaning it is ok
for this to happen, so this revises the test to accept it.